fix rmtree error action

+++ cameo/cleaner.py
@@ -9,6 +9,7 @@ This file is part of BSD license
 import shutil
 import os
 import stat
+import logging
 """
 清理不需要的資料
 """
@@ -17,11 +18,18 @@ class CleanerForINDIEGOGO:
     def __init__(self):
         self.strBasedir = "cameo_res\\source_html\\INDIEGOGO"
     
+    #rmtree 失敗時呼叫此方法
     def rmtreeOnError(self, funcRmtree, strPath, _):
+        logging.warning("rmtree 發生錯誤，嘗試修改檔案為可寫入模式，再進行刪除。")
         os.chmod(strPath, stat.S_IWRITE) #chmod to writeable
-        #funcRmtree(strPath)
-        os.remove(strPath)
+        if os.path.isdir(strPath):
+            #處理資料夾 (使用遞迴)
+            funcRmtree(strPath, onerror=self.rmtreeOnError)
+        else:
+            #處理檔案
+            os.remove(strPath)
     
+    #清除 _files 資料夾及其下所有內容
     def clean(self):
         for base, dirs, files in os.walk(self.strBasedir):
             for dir in dirs:
